Break apart rage normalizer to separate normalizers
