+++ src/analysis/retail/warrior/fury/modules/normalizers/RageNormalizers/constants.ts
